+++ src/main/java/com/garbagemule/MobArena/ArenaListener.java
@@ -99,6 +99,7 @@ import org.bukkit.potion.PotionEffectType;
 import org.bukkit.projectiles.ProjectileSource;
 
 import java.util.EnumSet;
+import java.util.HashMap;
 import java.util.HashSet;
 import java.util.LinkedList;
 import java.util.List;
@@ -118,6 +119,7 @@ public class ArenaListener
             protect;
     private boolean monsterExp,
             monsterInfight,
+            monsterTrack,
             pvpOn,               // pvp-enabled in config
             pvpEnabled = false,  // activated on first wave
             foodRegen,
@@ -131,6 +133,8 @@ public class ArenaListener
 
     private EnumSet<EntityType> excludeFromRetargeting;
 
+    private HashMap<Integer, Player> lastKnownPlayerTargets;
+
     public ArenaListener(Arena arena, MobArena plugin) {
         this.plugin = plugin;
         this.arena = arena;
@@ -143,6 +147,7 @@ public class ArenaListener
         this.protect          = s.getBoolean("protect",              true);
         this.monsterExp       = s.getBoolean("monster-exp",          false);
         this.monsterInfight   = s.getBoolean("monster-infight",      false);
+        this.monsterTrack     = s.getBoolean("monster-tracks-player",false);
         this.pvpOn            = s.getBoolean("pvp-enabled",          false);
         this.foodRegen        = s.getBoolean("food-regen",           false);
         this.lockFoodLevel    = s.getBoolean("lock-food-level",      true);
@@ -159,6 +164,8 @@ public class ArenaListener
             EntityType.ELDER_GUARDIAN,
             EntityType.GUARDIAN
         );
+
+        this.lastKnownPlayerTargets = new HashMap<>();
     }
 
     void pvpActivate() {
@@ -861,6 +868,18 @@ public class ArenaListener
             if (excludeFromRetargeting.contains(monster.getType())) {
                 return;
             }
+            // If monsterTrack is on, call overloaded getClosestPlayer
+            // that takes a monster's last known target as a parameter.
+            if (monsterTrack) {
+                Player lastTarget = null;
+                if (lastKnownPlayerTargets.containsKey(monster.getEntityId())) {
+                    lastTarget = lastKnownPlayerTargets.get(monster.getEntityId());
+                }
+                Player closestPlayer = MAUtils.getClosestPlayer(plugin, monster, arena, lastTarget);
+                lastKnownPlayerTargets.put(monster.getEntityId(), closestPlayer);
+                event.setTarget(closestPlayer);
+                return;
+            }
             event.setTarget(MAUtils.getClosestPlayer(plugin, monster, arena));
             return;
         }
@@ -876,6 +895,9 @@ public class ArenaListener
         // any foreign entity.
         if (!isArenaPlayer(target)) {
             event.setCancelled(true);
+        }
+        else if (monsterTrack) {
+            lastKnownPlayerTargets.put(monster.getEntityId(), (Player) target);
         }
     }